refactored, shifted to TS, clean up
