--- mongo/postv2.js
@@ -1,13 +1,0 @@
-const mongoose = require("mongoose");
-
-const Postv2 = new mongoose.Schema({
-  prompt: { type: String, required: true },
-  html: { type: String, required: true },
-  date: { type: Date, default: Date.now },
-  version: { type: Number, default: 2 },
-  author: { type: String },
-});
-
-const PostSchema = mongoose.model("Postv2", Postv2);
-
-module.exports = PostSchema;
