--- responses/newkey.json
@@ -1,21 +1,0 @@
-{
-  "id": "chatcmpl-70S7EJ18zd0MqDUCg3VncoM4KOy4m",
-  "object": "chat.completion",
-  "created": 1680342432,
-  "model": "gpt-3.5-turbo-0301",
-  "usage": {
-    "prompt_tokens": 134,
-    "completion_tokens": 735,
-    "total_tokens": 869
-  },
-  "choices": [
-    {
-      "message": {
-        "role": "assistant",
-        "content": "---starthtml---\n<div class=\"header\">\n  <!--Header Section-->\n  <h1>New Gaming Laptop Launch</h1>\n</div>\n<section class=\"about\">\n  <!--About Section-->\n  <h2>About</h2>\n  <p>Introducing our new Gaming Laptop! We've built it with the latest hardware to deliver an unparalleled gaming experience.</p>\n</section>\n<section class=\"specs\">\n  <!--Specs Section-->\n  <h2>Specifications</h2>\n  <ul>\n    <li>Processor: Intel Core i7</li>\n    <li>GPU: NVIDIA GeForce RTX 3080</li>\n    <li>RAM: 16GB DDR4</li>\n    <li>Storage: 1TB SSD</li>\n  </ul>\n  <div class=\"graph\">\n    <!--Graph Section-->\n    <h3>Performance Chart</h3>\n    <img src=\"performance_chart.png\" alt=\"Performance Chart\">\n  </div>\n</section>\n<section class=\"pricing\">\n  <!--Pricing Section-->\n  <h2>Pricing</h2>\n  <ul>\n    <li>Base Model: $1999</li>\n    <li>Customized Builds: Starting at $2199</li>\n  </ul>\n</section>\n<section class=\"location\">\n  <!--Location Section-->\n  <h2>Location</h2>\n  <iframe src=\"https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3150.389010078107!2d-122.40949398469848!3d37.78503498015061!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x808f7f54a7f887d3%3A0x2b44ff478688fc1e!2sApple%20Union%20Square!5e0!3m2!1sen!2sus!4v1635861848041!5m2!1sen!2sus\" width=\"600\" height=\"450\" style=\"border:0;\" allowfullscreen=\"\" loading=\"lazy\"></iframe>\n</section>\n---endhtml---\n\n---startcss---\n/* Base Styles */\nbody {\n  margin: 0;\n  padding: 0;\n  font-family: Arial, sans-serif;\n}\n\n/* Header Section */\n.header {\n  background-color: #1a1a1a;\n  color: #fff;\n  text-align: center;\n  padding: 20px;\n}\n\n/* About Section */\n.about {\n  background-color: #f2f2f2;\n  padding: 50px;\n}\n\n/* Specs Section */\n.specs {\n  background-color: #fff;\n  padding: 50px;\n}\n\n.graph {\n  margin-top: 25px;\n  text-align: center;\n}\n\n/* Pricing Section */\n.pricing {\n  background-color: #f2f2f2;\n  padding: 50px;\n}\n\n/* Location Section */\n.location {\n  background-color: #fff;\n  padding: 50px;\n  text-align: center;\n}\n\n/* Responsive Styles */\n@media (max-width: 600px) {\n  .specs,\n  .pricing {\n    padding: 25px;\n  }\n}\n---endcss---\n\n---startjs---\n// No JavaScript needed for this landing page\n---endjs---"
-      },
-      "finish_reason": "stop",
-      "index": 0
-    }
-  ]
-}
