--- responses/resp1.json
@@ -1,21 +1,0 @@
-{
-  "id": "chatcmpl-6zB75EAQHabCUnZG1wL5jZN5XbybE",
-  "object": "chat.completion",
-  "created": 1680038747,
-  "model": "gpt-3.5-turbo-0301",
-  "usage": {
-    "prompt_tokens": 109,
-    "completion_tokens": 1413,
-    "total_tokens": 1522
-  },
-  "choices": [
-    {
-      "message": {
-        "role": "assistant",
-        "content": "Here is an example of a shark hunting landing page without the HTML, HEAD, BODY, and SCRIPT tags as requested. The HTML code is wrapped between the `---starthtml---` and `---endhtml---` tags, the CSS code is wrapped between the `---startcss---` and `---endcss---` tags, and the JavaScript code is wrapped between the `---startjs---` and `---endjs---` tags.\n\n---starthtml---\n<div class=\"header\">\n  <h1>Shark Hunting Adventures</h1>\n  <nav>\n    <ul>\n      <li><a href=\"#\">Home</a></li>\n      <li><a href=\"#\">About Us</a></li>\n      <li><a href=\"#\">Gallery</a></li>\n      <li><a href=\"#\">Contact</a></li>\n    </ul>\n  </nav>\n</div>\n\n<main>\n  <section class=\"banner\">\n    <h2>Experience the Thrill of Shark Hunting</h2>\n    <p>Get up close and personal with some of the most fascinating creatures in the ocean.</p>\n    <a href=\"#\" class=\"cta-button\">Book Your Adventure</a>\n  </section>\n\n  <section class=\"features\">\n    <div class=\"feature\">\n      <img src=\"shark-fin.png\" alt=\"Shark Fin\">\n      <h3>Explore the Deep</h3>\n      <p>Discover the hidden wonders of the ocean and the amazing creatures that call it home.</p>\n    </div>\n\n    <div class=\"feature\">\n      <img src=\"shark-cage.png\" alt=\"Shark Cage\">\n      <h3>Safe and Secure</h3>\n      <p>Experience the thrill of shark hunting in a safe and controlled environment.</p>\n    </div>\n\n    <div class=\"feature\">\n      <img src=\"shark-trophy.png\" alt=\"Shark Trophy\">\n      <h3>Take Home a Trophy</h3>\n      <p>Show off your hunting prowess with a trophy from your adventure.</p>\n    </div>\n  </section>\n\n  <section class=\"testimonial\">\n    <blockquote>\"Shark hunting with these guys was an incredible experience. They know how to make you feel safe while still getting your adrenaline pumping. I would definitely recommend them to anyone looking for an adventure.\"</blockquote>\n    <cite>John D.</cite>\n  </section>\n</main>\n\n<footer>\n  <div class=\"footer-nav\">\n    <ul>\n      <li><a href=\"#\">Home</a></li>\n      <li><a href=\"#\">About Us</a></li>\n      <li><a href=\"#\">Gallery</a></li>\n      <li><a href=\"#\">Contact</a></li>\n    </ul>\n  </div>\n\n  <div class=\"footer-social\">\n    <a href=\"#\"><img src=\"facebook.png\" alt=\"Facebook\"></a>\n    <a href=\"#\"><img src=\"twitter.png\" alt=\"Twitter\"></a>\n    <a href=\"#\"><img src=\"instagram.png\" alt=\"Instagram\"></a>\n  </div>\n\n  <div class=\"footer-info\">\n    <p>Shark Hunting Adventures &copy; 2021</p>\n  </div>\n</footer>\n---endhtml---\n\n---startcss---\nbody {\n  margin: 0;\n  padding: 0;\n  font-family: Arial, sans-serif;\n}\n\n.header {\n  background-color: #333;\n  color: #fff;\n  display: flex;\n  justify-content: space-between;\n  align-items: center;\n  padding: 1rem;\n}\n\n.header h1 {\n  font-size: 2rem;\n  margin: 0;\n}\n\n.header nav ul {\n  list-style: none;\n  margin: 0;\n  padding: 0;\n}\n\n.header nav li {\n  display: inline-block;\n  margin-right: 1rem;\n}\n\n.header nav a {\n  color: #fff;\n  text-decoration: none;\n}\n\n.banner {\n  background-image: url('shark-banner.jpg');\n  background-size: cover;\n  height: 500px;\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n  align-items: center;\n  text-align: center;\n}\n\n.banner h2 {\n  font-size: 3rem;\n  color: #fff;\n  margin: 0;\n}\n\n.banner p {\n  font-size: 1.5rem;\n  color: #fff;\n  margin: 0 0 2rem 0;\n}\n\n.cta-button {\n  background-color: #f1c40f;\n  color: #333;\n  padding: 1rem 2rem;\n  border-radius: 2rem;\n  text-decoration: none;\n  font-weight: bold;\n}\n\n.features {\n  display: flex;\n  justify-content: space-between;\n  align-items: center;\n  padding: 2rem;\n}\n\n.feature {\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n  align-items: center;\n  text-align: center;\n  margin-right: 1rem;\n}\n\n.feature img {\n  width: 150px;\n  margin-bottom: 1rem;\n}\n\n.feature h3 {\n  font-size: 2rem;\n  margin: 0;\n}\n\n.feature p {\n  font-size: 1.2rem;\n}\n\n.testimonial {\n  background-color: #333;\n  color: #fff;\n  display: flex;\n  justify-content: space-between;\n  align-items: center;\n  padding: 2rem;\n}\n\n.testimonial blockquote {\n  font-size: 2rem;\n  margin: 0;\n}\n\n.testimonial cite {\n  font-size: 1.5rem;\n}\n\n.footer {\n  background-color: #333;\n  color: #fff;\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n  align-items: center;\n  padding-top: 2rem;\n  padding-bottom: 2rem;\n}\n\n.footer-nav ul {\n  list-style: none;\n  margin: 0;\n  padding: 0;\n  display: flex;\n  justify-content: center;\n  align-items: center;\n  margin-bottom: 1rem;\n}\n\n.footer-nav li {\n  display: inline-block;\n  margin-right: 1rem;\n}\n\n.footer-nav a {\n  color: #fff;\n  text-decoration: none;\n}\n\n.footer-social {\n  display: flex;\n  justify-content: center;\n  align-items: center;\n  margin-bottom: 1rem;\n}\n\n.footer-social img {\n  width: 30px;\n  margin-right: 1rem;\n}\n\n.footer-info {\n  font-size: 1.2rem;\n}\n---endcss---\n\n---startjs---\n// Code goes here\n---endjs---"
-      },
-      "finish_reason": "stop",
-      "index": 0
-    }
-  ]
-}
