--- responses/test.html
@@ -1,197 +1,0 @@
-"
-
-<!DOCTYPE html>
-<html>
-
-  <head>
-    <title>Singing Competition Landing Page</title>
-    
-    ---startcss---
-    /* Add your CSS code here */
-    body {
-      background-color: #f8f8f8;
-      font-family: Arial, sans-serif;
-      font-size: 16px;
-      color: #444444;
-    }
-    
-    .header {
-      background-color: #202020;
-      color: #ffffff;
-      padding: 20px;
-      text-align: center;
-    }
-    
-    .logo {
-      font-size: 30px;
-    }
-    
-    .subheader {
-      font-size: 20px;
-      margin-top: 30px;
-      text-align: center;
-    }
-    
-    .competition-description {
-      margin-top: 30px;
-      margin-bottom: 30px;
-      padding: 10px;
-      text-align: justify;
-    }
-    
-    .btn {
-      background-color: #ff2851;
-      color: #ffffff;
-      border: none;
-      padding: 10px 20px;
-      font-size: 16px;
-      cursor: pointer;
-      border-radius: 5px;
-    }
-    
-    .btn:hover {
-      background-color: #ffffff;
-      color: #ff2851;
-      border: 1px solid #ff2851;
-    }
-    
-    .footer {
-      background-color: #202020;
-      color: #ffffff;
-      padding: 20px;
-      text-align: center;
-      font-size: 14px;
-    }
-    
-    ---endcss---
-    
-  </head>
-
-  <body>
-    <div class="header">
-      <h1 class="logo">Singing Competition</h1>
-    </div>
-    
-    <div class="subheader">
-      <p>Are you the next singing superstar?</p>
-      <p>Enter our competition now and showcase your talent!</p>
-    </div>
-    
-    <div class="competition-description">
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec luctus finibus elementum. Sed ut sapien sodales, suscipit turpis vel, volutpat velit. Nam faucibus, eros at tempus tristique, mi tortor sodales ante, eu malesuada nulla ipsum ut sapien.</p>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec luctus finibus elementum. Sed ut sapien sodales, suscipit turpis vel, volutpat velit. Nam faucibus, eros at tempus tristique, mi tortor sodales ante, eu malesuada nulla ipsum ut sapien.</p>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
-    </div>
-    
-    <div class="btn-container">
-      <button class="btn">Enter now!</button>
-    </div>
-    
-    <div class="footer">
-      <p>Copyright © 2022 Singing Competition</p>
-    </div>
-    
-  </body>
-
-</html>
-
-"
-
-"---starthtml---
-
-<!DOCTYPE html>
-<html>
-
-  <head>
-    <title>Singing Competition Landing Page</title>
-    
-    ---startcss---
-    /* Add your CSS code here */
-    body {
-      background-color: #f8f8f8;
-      font-family: Arial, sans-serif;
-      font-size: 16px;
-      color: #444444;
-    }
-    
-    .header {
-      background-color: #202020;
-      color: #ffffff;
-      padding: 20px;
-      text-align: center;
-    }
-    
-    .logo {
-      font-size: 30px;
-    }
-    
-    .subheader {
-      font-size: 20px;
-      margin-top: 30px;
-      text-align: center;
-    }
-    
-    .competition-description {
-      margin-top: 30px;
-      margin-bottom: 30px;
-      padding: 10px;
-      text-align: justify;
-    }
-    
-    .btn {
-      background-color: #ff2851;
-      color: #ffffff;
-      border: none;
-      padding: 10px 20px;
-      font-size: 16px;
-      cursor: pointer;
-      border-radius: 5px;
-    }
-    
-    .btn:hover {
-      background-color: #ffffff;
-      color: #ff2851;
-      border: 1px solid #ff2851;
-    }
-    
-    .footer {
-      background-color: #202020;
-      color: #ffffff;
-      padding: 20px;
-      text-align: center;
-      font-size: 14px;
-    }
-    
-    ---endcss---
-    
-  </head>
-
-  <body>
-    <div class="header">
-      <h1 class="logo">Singing Competition</h1>
-    </div>
-    
-    <div class="subheader">
-      <p>Are you the next singing superstar?</p>
-      <p>Enter our competition now and showcase your talent!</p>
-    </div>
-    
-    <div class="competition-description">
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec luctus finibus elementum. Sed ut sapien sodales, suscipit turpis vel, volutpat velit. Nam faucibus, eros at tempus tristique, mi tortor sodales ante, eu malesuada nulla ipsum ut sapien.</p>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec luctus finibus elementum. Sed ut sapien sodales, suscipit turpis vel, volutpat velit. Nam faucibus, eros at tempus tristique, mi tortor sodales ante, eu malesuada nulla ipsum ut sapien.</p>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
-    </div>
-    
-    <div class="btn-container">
-      <button class="btn">Enter now!</button>
-    </div>
-    
-    <div class="footer">
-      <p>Copyright © 2022 Singing Competition</p>
-    </div>
-    
-  </body>
-
-</html>
-
----endhtml---"
