--- responses/testsend.json
@@ -1,21 +1,0 @@
-{
-  "id": "chatcmpl-6z8rQWYER7yOglIPvR6gkyY8jNW78",
-  "object": "chat.completion",
-  "created": 1680030088,
-  "model": "gpt-3.5-turbo-0301",
-  "usage": {
-    "prompt_tokens": 110,
-    "completion_tokens": 920,
-    "total_tokens": 1030
-  },
-  "choices": [
-    {
-      "message": {
-        "role": "assistant",
-        "content": "Sorry, as an AI language model, I can not provide visual output. However, I can provide you with the code for building a landing page for a house rental company.\n\nHere is a sample HTML and CSS code for a landing page. You can customize it according to your requirements.\n\nHTML:\n\n```\n---starthtml---\n<header>\n  <nav>\n    <ul>\n      <li><a href=\"#\">Home</a></li>\n      <li><a href=\"#\">About Us</a></li>\n      <li><a href=\"#\">Contact Us</a></li>\n    </ul>\n  </nav>\n</header>\n\n<main>\n  <section class=\"banner\">\n    <h1>Welcome to XYZ House Rental Company</h1>\n    <p>Explore our vast range of rental properties and find your dream home</p>\n    <a href=\"#\" class=\"btn\">Browse Properties</a>\n  </section>\n\n  <section class=\"properties\">\n    <h2>Featured Properties</h2>\n    <div class=\"property\">\n      <div class=\"img-wrapper\">\n        <img src=\"property-1.jpg\" alt=\"Property 1\">\n      </div>\n      <h3>Property Name</h3>\n      <p>Location</p>\n      <a href=\"#\" class=\"btn\">View Details</a>\n    </div>\n    <div class=\"property\">\n      <div class=\"img-wrapper\">\n        <img src=\"property-2.jpg\" alt=\"Property 2\">\n      </div>\n      <h3>Property Name</h3>\n      <p>Location</p>\n      <a href=\"#\" class=\"btn\">View Details</a>\n    </div>\n    <div class=\"property\">\n      <div class=\"img-wrapper\">\n        <img src=\"property-3.jpg\" alt=\"Property 3\">\n      </div>\n      <h3>Property Name</h3>\n      <p>Location</p>\n      <a href=\"#\" class=\"btn\">View Details</a>\n    </div>\n  </section>\n</main>\n---endhtml---\n```\n\nCSS:\n\n```\n---startcss---\n* {\n  margin: 0;\n  padding: 0;\n  box-sizing: border-box;\n}\n\nheader {\n  background-color: #fff;\n  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);\n  padding: 1rem;\n}\n\nnav ul {\n  list-style: none;\n  display: flex;\n  justify-content: space-around;\n}\n\nnav a {\n  text-decoration: none;\n  color: #333;\n  font-weight: bold;\n  font-size: 1.2rem;\n  padding: 0.5rem;\n}\n\n.banner {\n  background-image: url('banner.jpg');\n  background-size: cover;\n  height: calc(100vh - 80px);\n  display: flex;\n  flex-direction: column;\n  justify-content: center;\n  align-items: center;\n  text-align: center;\n  color: #fff;\n}\n\n.banner h1 {\n  font-size: 3rem;\n  margin-bottom: 1rem;\n}\n\n.banner p {\n  font-size: 1.2rem;\n  margin-bottom: 2rem;\n}\n\n.properties {\n  margin: 2rem 0;\n}\n\n.properties h2 {\n  font-size: 2rem;\n  text-align: center;\n  margin-bottom: 2rem;\n}\n\n.property {\n  margin: 2rem;\n  border: 1px solid #ccc;\n  padding: 1rem;\n  text-align: center;\n}\n\n.property h3 {\n  font-size: 1.5rem;\n  margin: 1rem 0;\n}\n\n.property p {\n  font-size: 1.2rem;\n  margin: 1rem 0;\n}\n\n.property .img-wrapper {\n  height: 200px;\n  overflow: hidden;\n  margin-bottom: 1rem;\n}\n\n.property img {\n  width: 100%;\n  height: auto;\n}\n\n.btn {\n  display: inline-block;\n  background-color: #333;\n  color: #fff;\n  text-decoration: none;\n  font-weight: bold;\n  padding: 0.5rem 1rem;\n  border-radius: 4px;\n  transition: all 0.2s;\n}\n\n.btn:hover {\n  background-color: #666;\n}\n---endcss---\n```"
-      },
-      "finish_reason": "stop",
-      "index": 0
-    }
-  ]
-}
