--- responses/testsend2.json
@@ -1,21 +1,0 @@
-{
-  "id": "chatcmpl-6zAPIWYWx4DvuKCjbter5JyfzOorK",
-  "object": "chat.completion",
-  "created": 1680036032,
-  "model": "gpt-3.5-turbo-0301",
-  "usage": {
-    "prompt_tokens": 111,
-    "completion_tokens": 1528,
-    "total_tokens": 1639
-  },
-  "choices": [
-    {
-      "message": {
-        "role": "assistant",
-        "content": "Sorry, as an AI language model, I cannot generate visual designs. However, I can provide you with sample HTML code for a fitness gym landing page with vibrant colors:\n\n```\n---starthtml---\n<head>\n\t<title>Fitness Gym Landing Page</title>\n\t<style>\n\t\t/* CSS for the landing page */\n\t\tbody {\n\t\t\tbackground-color: #f5f5f5;\n\t\t\tfont-family: Arial, sans-serif;\n\t\t\tmargin: 0;\n\t\t\tpadding: 0;\n\t\t}\n\n\t\theader {\n\t\t\tbackground-color: #FF8C00;\n\t\t\tcolor: #fff;\n\t\t\tpadding: 20px;\n\t\t\ttext-align: center;\n\t\t}\n\n\t\th1 {\n\t\t\tfont-size: 3rem;\n\t\t\tmargin: 0;\n\t\t}\n\n\t\tnav {\n\t\t\tbackground-color: #fff;\n\t\t\tpadding: 10px;\n\t\t\ttext-align: center;\n\t\t}\n\n\t\tnav ul {\n\t\t\tlist-style-type: none;\n\t\t\tmargin: 0;\n\t\t\tpadding: 0;\n\t\t}\n\n\t\tnav li {\n\t\t\tdisplay: inline;\n\t\t\tmargin: 20px;\n\t\t}\n\n\t\tnav a {\n\t\t\tcolor: #FF8C00;\n\t\t\ttext-decoration: none;\n\t\t\tfont-size: 1.5rem;\n\t\t}\n\n\t\tsection {\n\t\t\tbackground-color: #fff;\n\t\t\tdisplay: flex;\n\t\t\tjustify-content: center;\n\t\t\tpadding: 50px;\n\t\t}\n\n\t\tsection h2 {\n\t\t\tcolor: #FF8C00;\n\t\t\tfont-size: 2rem;\n\t\t\tmargin-bottom: 30px;\n\t\t\ttext-transform: uppercase;\n\t\t\tletter-spacing: 2px;\n\t\t}\n\n\t\t.card {\n\t\t\tbackground-color: #f5f5f5;\n\t\t\tbox-shadow: 0 0 10px rgba(0,0,0,0.2);\n\t\t\tdisplay: flex;\n\t\t\tflex-direction: column;\n\t\t\tpadding: 50px;\n\t\t\ttext-align: center;\n\t\t\twidth: 300px;\n\t\t\theight: 400px;\n\t\t\tmargin: 20px;\n\t\t\ttransition: transform 0.3s;\n\t\t}\n\n\t\t.card:hover {\n\t\t\ttransform: scale(1.1);\n\t\t}\n\n\t\t.card h3 {\n\t\t\tmargin-top: 0;\n\t\t}\n\n\t\t.card img {\n\t\t\theight: 200px;\n\t\t\tmargin: 20px;\n\t\t}\n\n\t\t.card p {\n\t\t\tmargin: 20px;\n\t\t}\n\n\t\t.call-to-action {\n\t\t\tbackground-color: #FF8C00;\n\t\t\tcolor: #fff;\n\t\t\tdisplay: flex;\n\t\t\tflex-wrap: wrap;\n\t\t\tjustify-content: space-between;\n\t\t\tpadding: 50px;\n\t\t}\n\n\t\t.call-to-action h2 {\n\t\t\tfont-size: 2rem;\n\t\t\tmargin-top: 0;\n\t\t\tmargin-bottom: 50px;\n\t\t\twidth: 100%;\n\t\t\ttext-align: center;\n\t\t}\n\n\t\tform {\n\t\t\tdisplay: flex;\n\t\t\tflex-wrap: wrap;\n\t\t\tjustify-content: center;\n\t\t\twidth: 100%;\n\t\t}\n\n\t\tinput[type=\"email\"] {\n\t\t\tborder-radius: 5px;\n\t\t\tborder: none;\n\t\t\tmargin: 10px;\n\t\t\tpadding: 10px;\n\t\t\twidth: 60%;\n\t\t\tfont-size: 1.2rem;\n\t\t}\n\n\t\tinput[type=\"submit\"] {\n\t\t\tbackground-color: #fff;\n\t\t\tborder: 2px solid #fff;\n\t\t\tborder-radius: 5px;\n\t\t\tcolor: #FF8C00;\n\t\t\tcursor: pointer;\n\t\t\tfont-size: 1.2rem;\n\t\t\tmargin: 10px;\n\t\t\tpadding: 10px;\n\t\t\ttransition: background-color 0.2s;\n\t\t}\n\n\t\tinput[type=\"submit\"]:hover {\n\t\t\tbackground-color: #FF8C00;\n\t\t\tcolor: #fff;\n\t\t}\n\n\t\tfooter {\n\t\t\tbackground-color: #FF8C00;\n\t\t\tcolor: #fff;\n\t\t\tpadding: 20px;\n\t\t\ttext-align: center;\n\t\t}\n\n\t\tfooter p {\n\t\t\tmargin: 0;\n\t\t}\n\t</style>\n</head>\n<body>\n\t<header>\n\t\t<h1>Fitness Gym</h1>\n\t</header>\n\t<nav>\n\t\t<ul>\n\t\t\t<li><a href=\"#what-we-offer\">What We Offer</a></li>\n\t\t\t<li><a href=\"#testimonials\">Testimonials</a></li>\n\t\t\t<li><a href=\"#contact-us\">Contact Us</a></li>\n\t\t</ul>\n\t</nav>\n\t<section id=\"what-we-offer\">\n\t\t<h2>What We Offer</h2>\n\t\t<div class=\"card\">\n\t\t\t<img src=\"https://via.placeholder.com/200x200\" alt=\"Card Image\">\n\t\t\t<h3>Card One</h3>\n\t\t\t<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus sit amet tortor ac nulla facilisis interdum.</p>\n\t\t</div>\n\t\t<div class=\"card\">\n\t\t\t<img src=\"https://via.placeholder.com/200x200\" alt=\"Card Image\">\n\t\t\t<h3>Card Two</h3>\n\t\t\t<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus sit amet tortor ac nulla facilisis interdum.</p>\n\t\t</div>\n\t\t<div class=\"card\">\n\t\t\t<img src=\"https://via.placeholder.com/200x200\" alt=\"Card Image\">\n\t\t\t<h3>Card Three</h3>\n\t\t\t<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus sit amet tortor ac nulla facilisis interdum.</p>\n\t\t</div>\n\t</section>\n\t<section id=\"testimonials\">\n\t\t<h2>Testimonials</h2>\n\t\t<div class=\"card\">\n\t\t\t<p>\"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus sit amet tortor ac nulla facilisis interdum.\"</p>\n\t\t\t<p>- John Doe</p>\n\t\t</div>\n\t\t<div class=\"card\">\n\t\t\t<p>\"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus sit amet tortor ac nulla facilisis interdum.\"</p>\n\t\t\t<p>- Jane Doe</p>\n\t\t</div>\n\t\t<div class=\"card\">\n\t\t\t<p>\"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus sit amet tortor ac nulla facilisis interdum.\"</p>\n\t\t\t<p>- Jack Doe</p>\n\t\t</div>\n\t</section>\n\t<section id=\"contact-us\">\n\t\t<div class=\"call-to-action\">\n\t\t\t<h2>Join Our Gym Today!</h2>\n\t\t\t<form>\n\t\t\t\t<input type=\"email\" placeholder=\"Enter Your Email\">\n\t\t\t\t<input type=\"submit\" value=\"Subscribe\">\n\t\t\t</form>\n\t\t</div>\n\t</section>\n\t<footer>\n\t\t<p>&copy; 2021 Fitness Gym</p>\n\t</footer>\n</body>\n---endhtml---\n```\n\nThis HTML code includes a header with a vibrant orange background color, navigation bar, sections for \"What We Offer\" and \"Testimonials\" with cards arranged in a flexbox, a section for a call-to-action to join the gym with a form, and a footer."
-      },
-      "finish_reason": "stop",
-      "index": 0
-    }
-  ]
-}
