+++ src/constants/index.ts
@@ -1,0 +1,5 @@
+export enum TogetherAIModels {
+  Mixtral8x7B = "mistralai/Mixtral-8x7B-Instruct-v0.1",
+  OpenHermes2p5Mistral7B = "teknium/OpenHermes-2p5-Mistral-7B",
+  NousHermes2Mixtral8x7BDPO = "NousResearch/Nous-Hermes-2-Mixtral-8x7B-DPO",
+}
