+++ src/routes/testRoutes.ts
@@ -1,5 +1,8 @@
-const express = require("express");
-const router = express.Router();
+import { Router } from "express";
+const router = Router();
+
+import ApiV2Controller from "../controllers/ApiV2Controller";
+
 const {
   betaCompletion,
   createPost,
@@ -8,7 +11,7 @@ const {
   getPostbyId,
   getPaginatedGenerations,
   getImages,
-} = require("../controllers/testController");
+} = ApiV2Controller;
 
 router.post("/test", betaCompletion);
 router.post("/dummy", dummyCompletion);
@@ -18,4 +21,5 @@ router.put("/update/:id", updatePost);
 router.get("/postv2/:id", getPostbyId);
 router.get("/betapagination", getPaginatedGenerations);
 router.get("/searchImages", getImages);
-module.exports = router;
+
+export default router;
